requirements.txt
tidevice==0.7.14
PyMySQL==1.0.2
jsonpath==0.82
xlwt==1.3.0
requests==2.27.1
numpy==1.22.4
py-ios-device==2.2.3.5
pyinstaller